refactor/#92 : 인증 검증을 통한 인증 로직을 분리한다
