+++ src/types/location.ts
@@ -1,0 +1,4 @@
+export type Location = {
+  latitude: number;
+  longitude: number;
+};
